requirements.txt
boto3>=1.26.0
PyYAML>=6.0
tabulate>=0.9.0
hypothesis>=6.0.0
pytest>=7.0.0
